added BeerTableView
